Akka Extension for configurating led status tracking
