+++ exercise_001_cluster_base/src/main/resources/application.conf
@@ -54,7 +54,7 @@ akka {
 
     artery {
       canonical {
-        hostname = "127.0.0.1"
+        hostname = ${cluster-node-configuration.node-hostname}
         port = 2550
       }
     }
